Minor changes to get code running in Python2

+++ mmdet/models/anchor_heads/anchor_head.py
@@ -129,7 +129,7 @@ class AnchorHead(nn.Module):
                     bbox_targets, bbox_weights, num_total_samples, cfg):
         # classification loss
         if self.use_sigmoid_cls:
-            labels = labels.reshape(-1, self.cls_out_channels)
+            labels = labels.reshape(-1, self.cls_out_channels).float()
             label_weights = label_weights.reshape(-1, self.cls_out_channels)
         else:
             labels = labels.reshape(-1)
